adding generate and score fn to utils
