+++ trl/trainer/utils.py
@@ -20,8 +20,10 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
+from accelerate import Accelerator
 from torch.nn.utils.rnn import pad_sequence
-from torch.utils.data import IterableDataset
+from torch.utils.data import DataLoader, IterableDataset
+from tqdm import tqdm
 from transformers import DataCollatorForLanguageModeling, PreTrainedModel, PreTrainedTokenizerBase, TrainerCallback
 
 
@@ -761,3 +763,127 @@ def neftune_forward(self, input: torch.Tensor):
         embeddings = embeddings + torch.zeros_like(embeddings).uniform_(-mag_norm, mag_norm)
 
     return embeddings
+
+
+@torch.no_grad()
+def generate(
+    model: PreTrainedModel,
+    dataloader: DataLoader,
+    tokenizer: PreTrainedTokenizerBase,
+    accelerator: Accelerator,
+    **generation_kwargs,
+) -> Tuple[List[str], List[str]]:
+    """Generate samples using model using specified generation_kwargs.
+
+    Args:
+        model: the model used to score samples.
+        dataloader: the dataloader containing batches of elements from the generated dataset.
+        tokenizer: the tokenizer.
+        accelerator: the accelerator object.
+
+    Returns:
+        all_prompts: the prompts.
+        all_responses: the responses generated by the model.
+    """
+
+    all_predictions = []
+    all_prompts = []
+    pbar = tqdm(total=len(dataloader), disable=not accelerator.is_local_main_process)
+
+    for batch in dataloader:
+        sequence_length = batch["input_ids"].shape[1]
+
+        all_tokens = accelerator.unwrap_model(model).generate(
+            batch["input_ids"],
+            attention_mask=batch["attention_mask"],
+            **generation_kwargs,
+        )
+
+        generated_tokens = all_tokens[:, sequence_length:]
+        prompt_tokens = all_tokens[:, :sequence_length]
+
+        generated_tokens = accelerator.pad_across_processes(generated_tokens, dim=1, pad_index=tokenizer.pad_token_id)
+        prompt_tokens = accelerator.pad_across_processes(prompt_tokens, dim=1, pad_index=tokenizer.pad_token_id)
+
+        generated_tokens = accelerator.gather(generated_tokens)
+        generated_tokens = generated_tokens.cpu()
+        prompt_tokens = accelerator.gather(prompt_tokens)
+        prompt_tokens = prompt_tokens.cpu()
+
+        if isinstance(generated_tokens, tuple):
+            generated_tokens = generated_tokens[0]
+            prompt_tokens = prompt_tokens[0]
+
+        all_predictions.extend(generated_tokens)
+        all_prompts.extend(prompt_tokens)
+        pbar.update(1)
+
+    accelerator.wait_for_everyone()
+
+    all_predictions = tokenizer.batch_decode(all_predictions, skip_special_tokens=True)
+    all_prompts = tokenizer.batch_decode(all_prompts, skip_special_tokens=True)
+
+    return all_prompts, all_predictions
+
+
+@torch.no_grad()
+def compute_reward_score(model: PreTrainedModel, dataloader: DataLoader, accelerator: Accelerator) -> List[float]:
+    """Score the generated dataset based on a reward model.
+
+    Args:
+        model: the model used to score samples.
+        dataloader: the dataloader containing batches of elements from the generated dataset.
+        accelerator: the accelerator object.
+
+    Returns:
+        rewards: rewards assigned to each sample of the generated dataset.
+    """
+
+    rewards = []
+    pbar = tqdm(total=len(dataloader), disable=not accelerator.is_local_main_process)
+
+    for batch in dataloader:
+        scores = model(**batch).logits.squeeze(1)
+        scores = accelerator.gather(scores)
+        rewards.extend(scores)
+        pbar.update(1)
+
+    rewards = [reward.item() for reward in rewards]
+
+    return rewards
+
+
+# modified from https://arxiv.org/pdf/2309.06657.pdf
+def conduct_rejection_sampling(
+    response_candidates: List[str], response_rewards: List[float], num_samples: int, beta: float
+) -> Tuple[List[str], List[float]]:
+    """Conducts statistical rejection sampling.
+
+    Args:
+        response_candidates: response candidates from sft policy
+        response_rewards: response rewards.
+        num_samples: number of samples to sub-sample.
+        beta: beta parameter in KL-constrained reward maximization objective.
+
+    Returns:
+        accepted: Accepted rejection sampled sequences from the optimal policy.
+        rewards: the rewards associated to the accepted samples.
+    """
+    candidates = {c: r for c, r in zip(response_candidates, response_rewards)}
+    accepted = []
+    rewards = []
+    while len(accepted) < num_samples:
+        max_reward = max(candidates.values())
+        to_remove = []
+        for c, r in candidates.items():
+            u = np.random.uniform()
+            if u >= np.exp((r - max_reward) / beta):
+                continue
+            accepted.append(c)
+            rewards.append(r)
+            to_remove.append(c)
+            if len(accepted) == num_samples:
+                break
+        for c in to_remove:
+            candidates.pop(c)
+    return accepted, rewards
